backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
anthropic>=0.26.0
openai>=1.50.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
pydantic>=2.10.0
google-api-python-client>=2.100.0
google-auth>=2.27.0

# ── NEW: Production additions ────────────────────────────────────
sqlalchemy>=2.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# ── Testing ──────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # For FastAPI TestClient

